--- index.html
@@ -1,14 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="icon" type="image/png" href="src/assets/logo.png" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>ViZarrStores</title>
-  </head>
-  <body>
-    <div id="root"></div>
-    <div id="overlays"></div>
-    <script type="module" src="/src/main.tsx"></script>
-  </body>
-</html>
