+++ package_old.json
@@ -1,0 +1,60 @@
+{
+  "name": "ViZarrStores",
+  "description": "A browser based visualization toolkit for Zarr Stores",
+  "license": "Apache-2.0",
+  "author": {
+    "name": "Lazaro Alonso & Jeran Poehls",
+    "email": "lazarus.alon@gmail.com",
+    "url": "https://lazarusa.github.io"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/EarthyScience/ViZarrStores.git"
+  },
+  "private": true,
+  "version": "0.0.1",
+  "type": "module",
+  "packageManager": "pnpm@10.8.1",
+  "scripts": {
+    "prChecks": "pnpm lint && pnpm buildRepo",
+    "buildRepo": "pnpm rmBuild && pnpm build",
+    "rmBuild": "pnpm -r exec rm -rf dist",
+    "test": "vitest run",
+    "dev": "vite --host",
+    "build": "tsc && vite build",
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 20",
+    "preview": "vite preview"
+  },
+  "dependencies": {
+    "@react-three/drei": "^10.0.6",
+    "@react-three/fiber": "^9.1.2",
+    "@types/three": "^0.175.0",
+    "jotai": "^2.10.0",
+    "js-colormaps-es": "^0.0.5",
+    "leva": "^0.10.0",
+    "quick-lru": "^7.0.1",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "three": "^0.175.0",
+    "vite-plugin-glsl": "^1.4.0",
+    "vitest": "^3.1.1",
+    "zarrita": "^0.5.1"
+  },
+  "devDependencies": {
+    "@eslint/js": "^9.24.0",
+    "@types/node": "^22.14.1",
+    "@types/react": "^19.1.2",
+    "@types/react-dom": "^19.1.2",
+    "@typescript-eslint/eslint-plugin": "^8.30.1",
+    "@typescript-eslint/parser": "^8.30.1",
+    "@vitejs/plugin-react": "^4.4.0",
+    "eslint": "^8.57.1",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-plugin-react-refresh": "^0.4.19",
+    "globals": "^16.0.0",
+    "typescript": "^5.6.2",
+    "typescript-eslint": "^8.30.1",
+    "vite": "^6.2.5"
+  }
+}
