--- src/App.css
@@ -1,222 +1,0 @@
-html, body, #root {
-    height: 100%;
-    margin: 0;
-    padding: 0;
-  }
-.pane {
-  position: absolute;
-  top: 20%;
-  left: 3%;
-  z-index: 10;
-}
-
-.panePlugin {
-  position: absolute;
-  top: 40px;
-  right: 2px;
-  z-index: 10;
-}
-
-.description {
-  position: absolute;
-  top: 44px;
-  left: 2px;
-  width: fit-content;
-  max-width: 80px;
-  padding: 0.25rem;
-  border-radius: 0.5rem;
-  background: #000000ab;
-  word-wrap: break-word;
-  overflow-wrap: break-word;
-  font-size: 12px;
-  line-height: 1.5;
-  z-index: 5;
-  color: #dfdcdc;
-}
-
-.description:hover {
-  max-width: 300px;
-}
-
-.hiddenMeta {
-  max-height: 0;
-  max-width: 265px;
-  opacity: 0;
-  visibility: hidden;
-  overflow: hidden;
-  transition: max-height 0.3s ease-in-out, opacity 0.3s ease-in-out;
-}
-.description:hover .hiddenMeta {
-  visibility: visible;
-  max-height: 300px;
-  padding: 0.2rem;
-  opacity: 1;
-  overflow: auto;
-}
-
-.metadata-content {
-  max-height: 0;
-  opacity: 0;
-  visibility: hidden;
-  overflow: hidden;
-  transition: max-height 0.3s ease-in-out, opacity 0.3s ease-in-out, visibility 0.3s ease-in-out;
-}
-
-.metadata-header:hover + .metadata-content {
-  max-height: 300px;
-  opacity: 1;
-  visibility: visible;
-}
-
-
-.metadata-header {
-  cursor: pointer;
-  padding-top: 0px;
-  padding-bottom: 2px;
-  padding-left: 5px;
-  padding-right: 5px;
-  margin: 0;
-}
-#colorbar-container {
-  position: fixed;
-  top: 4px;
-  left: 50%;
-  transform: translateX(-50%);
-  width: 360px;
-  height: 30px;
-  background-color: #000000aa;
-  border-radius: 4px;
-  display: grid;
-  grid-template-columns: auto 65% auto;
-  grid-template-rows: 1fr 1fr;
-  align-items: center;
-  padding: 0 10px;
-  gap: 2px;
-}
-
-.left-label {
-  grid-column: 1;
-  grid-row: 1 / span 2;
-  justify-self: start;
-  color: #f5c024;
-  font-size: 13px;
-  font-weight: bold;
-  transform: translateX(4px) translateY(-8px);
-  max-width: 50px;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-
-  &:hover {
-    overflow-x: auto;
-    scrollbar-width: none;  /* Firefox */
-    -ms-overflow-style: none;  /* IE and Edge */
-    white-space: nowrap;
-    max-width: 150px;
-    background-color: #000000aa;
-    border-radius: 3px;
-    z-index: 11;
-  }
-  /* Hide scrollbar for Chrome, Safari and Opera */
-  &::-webkit-scrollbar {
-    display: none;
-  }
-}
-
-.right-label {
-  grid-column: 3;
-  grid-row: 1 / span 2;
-  justify-self: end;
-  color: #f5c024;
-  font-size: 13px;
-  font-weight: bold;
-  transform: translateX(-4px) translateY(-8px);
-  max-width: 40px;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-
-  &:hover {
-    overflow-x: auto;
-    scrollbar-width: none;  /* Firefox */
-    -ms-overflow-style: none;  /* IE and Edge */
-    white-space: nowrap;
-    max-width: 150px;
-    background-color: #000000aa;
-    border-radius: 3px;
-    z-index: 11;
-  }
-  /* Hide scrollbar for Chrome, Safari and Opera */
-  &::-webkit-scrollbar {
-    display: none;
-  }
-}
-#colorbar {
-  grid-column: 2;
-  grid-row: 1;
-  width: 100%;
-  height: 12px;
-  background-color: #ffffff;
-}
-
-#ticks {
-  grid-column: 2;
-  grid-row: 2;
-  position: relative;
-  width: 100%;
-  height: 30px;
-}
-
-.tick {
-  position: absolute;
-  width: 1px;
-  height: 14.5px;
-  top: -14px;
-  background-color: var(--secondary-text-color);
-}
-
-.tick-label {
-  position: absolute;
-  padding-top: 0px;
-  font-size: 12px;
-  transform: translateX(-50%);
-  text-align: center;
-}
-
-@media (max-width: 896px) {
-  .panePlugin {
-    top: 44px;
-  }
-  .description {
-    width: fit-content;
-    max-width: 80px;
-    left: 2px;
-    padding: 0.25rem;
-  }
-  .description:hover {
-    max-width: 240px;
-  }
-
-  .description:hover .hiddenMeta {
-    max-height: 300px;
-    max-width: 240px;
-    overflow: auto;
-  }
-  #colorbar-container {
-    width: 65%;
-    padding: 0 5px;
-  }
-}
-
-.loading{
-  position: absolute;
-  left: 50%;
-  top: 50%;
-  background-color: rgba(255, 255, 255, 0.596);
-  z-index: 10;
-  padding: 20px;
-  font-size: 48px;
-  color: #222222;
-  border-radius: 20px;
-  transform: translate(-50%,-50%);
-}
