--- src/App.tsx
@@ -1,15 +1,0 @@
-import { Navbar, Footer, CanvasGeometry, } from '@/components/UI';
-import Overlays from "@/components/UI/Overlays";
-import './App.css'
-export function App() {
-  return (
-    <>
-    <Overlays />
-    <Navbar />
-    <Footer />
-    <CanvasGeometry />
-    </>
-  )
-}
-
-export default App
