+++ src/app/globals.css
@@ -1,0 +1,61 @@
+@import "tailwindcss";
+
+:root {
+  --background: #f5f5f5;
+  --foreground: #171717;
+  --background-modal: #d0cdcdcc;
+  --modal-shadow: #f5f1f1aa;
+  --modal-gradient: #cccaca;
+  --modal-p: #08000080;
+  --accent-1: black;
+}
+/* Light Theme */
+[data-theme='light'] {
+  --background: #FAFAFA;
+  --foreground: #171717;
+  --background-modal: #d0cdcdcc;
+  --modal-shadow: #f5f1f1aa;
+  --modal-gradient: #cccaca;
+  --modal-p: #08000080;
+  --accent-1: black;
+  --background-funding: #F1F2F4;
+  --toggle-hover: rgb(127, 128, 132);
+  --accent-1: black;
+
+}
+
+/* Dark Theme */
+[data-theme='dark'] {
+  --background: rgb(26, 26, 26);
+  --foreground: #ededed;
+  --background-modal: rgba(17, 16, 16, 0.667);
+  --modal-shadow: #000000aa;
+  --modal-gradient: #0a0a0a;
+  --modal-p: #ffffff80;
+  --background-funding: rgb(32, 33, 38);
+  --toggle-hover: rgb(239, 198, 95);
+  --accent-1: gold;
+}
+
+body {
+  background: var(--background);
+  color: var(--foreground);
+  font-family: Arial, Helvetica, sans-serif;
+}
+
+.canvas {
+  display: block;
+  height: 100vh;
+  width: 100vw;
+  margin: 0;
+  padding: 0;
+}
+ul {
+  list-style-type: disc;
+  padding-left: 2rem;
+  margin: 0.5rem 0;
+}
+
+li {
+  padding: 0.25rem 0;
+}
