+++ src/app/page.tsx
@@ -1,0 +1,4 @@
+import CanvasGeometry from "@/components/CanvasGeometry"
+export default function Home() {
+  return <CanvasGeometry />
+}
