+++ src/components/UI/AboutButton.tsx
@@ -1,5 +1,9 @@
+'use client';
+
 import { useSetAtom } from "jotai";
-import { uiAtom } from "@/state";
+import { uiAtom } from "@/components/ui";
+
+import "./css/AboutButton.css";
   const AboutButton = () => {
     const setUi = useSetAtom(uiAtom);
     return (
