--- src/components/UI/AboutModal.css
@@ -1,57 +1,0 @@
-.modal {
-  position: fixed;
-  top: 0;
-  left: 0px;
-  z-index: 10;
-  width: 100vw;
-  height: 100vh;
-  background-color: #000000aa;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  overflow: hidden; /* Ensure no scrollbars appear on the modal container */
-}
-
-.modal-content {
-  background-color: #8ec5fc;
-  background-image: linear-gradient(62deg, #191a2a 0%, #12100e 100%);
-  position: relative;
-  max-width: 25rem;
-  max-height: 80vh; /* Limit the max-height to 80% of viewport height */
-  margin: 0 auto;
-  padding: 1rem;
-  border-radius: 0.5rem;
-  box-shadow: 0 0 1em #000000aa;
-  overflow-y: auto; /* Enable vertical scrolling */
-}
-
-.modal-content h2 {
-  color: orangered;
-}
-
-.modal-content p {
-  color: hsla(0, 0%, 100%, 0.5);
-  font-size: 14px;
-}
-
-.modal-content ul {
-  color: hsla(0, 0%, 100%, 0.5);
-  font-size: 14px;
-}
-
-.close-btn {
-  position: absolute;
-  top: 0.5rem;
-  right: 0.5rem;
-  color: rgb(70, 70, 70);
-}
-
-.modal-content a {
-  color: var(--accent-color1);
-  text-align: center;
-}
-
-.modal-content h1 {
-  color: orangered;
-  font-size: 1.5rem;
-}
