--- src/components/UI/Footer.css
@@ -1,90 +1,0 @@
-.footer {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    color: var(--secondary-text-color);
-    background-color: #000000aa;
-    z-index: 1;
-    font-size: 13.5px;
-    position: fixed;
-    bottom: 0;
-    left: 0;
-    padding: 0.25rem 0.5rem;
-    width: 100%;
-}
-
-.footer p {
-    margin: 0;
-    flex: 1;
-    text-align: center;
-}
-.footer span {
-    margin: 0;
-    text-align: center;
-}
-.footer a {
-    position: relative;
-    text-decoration: none;
-    color: rgb(241, 233, 233);
-    padding: 0 0.25rem;
-}
-
-.footer a::after {
-    content: '';
-    position: absolute;
-    width: 0;
-    height: 1px;
-    bottom: 0;
-    left: 0;
-    background-color: rgb(205, 193, 193);
-    transition: width 0.3s;
-}
-
-.footer a:hover::after {
-    width: 100%;
-}
-
-.large-screen-logo {
-    display: block;
-}
-
-.small-screen-logos {
-    display: none;
-}
-.expandable-text {
-    display: flex;
-    flex: 1;
-    white-space: nowrap;
-}
-/* Responsive design for screens below 600px */
-@media (max-width: 600px) {
-    .footer {
-        display: grid;
-        grid-template-columns: auto 1fr auto;
-        align-items: center;
-        text-align: center;
-    }
-
-    .footer p {
-        text-align: center;
-        flex: unset;
-        margin: 0;
-    }
-
-    .large-screen-logo {
-        display: none;
-    }
-
-    .small-screen-logos {
-        display: flex;
-        justify-content: space-between;
-    }
-    .expandable-text {
-        display: unset;
-        margin-left: 0;
-        margin-right: 0;
-    }
-    .footer img {
-        max-height: 40px;
-    }
-}
