+++ src/components/UI/Footer.tsx
@@ -1,29 +1,29 @@
+import Image from "next/image";
 import { logoBGC_MPI, logoBGC, logoMPI } from "@/assets/index";
 
-import './Footer.css';
+import './css/Footer.css';
 
 const Footer = () => (
   <div className="footer">
     <div className="large-screen-logo">
       <a href="https://www.bgc-jena.mpg.de/en/bgi/mdi" target="_blank" rel="noreferrer">
-        <img src={logoBGC_MPI} alt="logoMPI" height={35}/>
+        <Image src={logoBGC_MPI} alt="logoMPI" height={32}/>
       </a>
     </div>
-    <div className="small-screen-logos">
+    <div className="small-screen-logo">
       <a href="https://www.bgc-jena.mpg.de/en/bgi/mdi" target="_blank" rel="noreferrer">
-        <img src={logoBGC} alt="logoBGC" height={35}/>
+        <Image src={logoBGC} alt="logoBGC" height={32}/>
       </a>
     </div>
     <div className="expandable-text">
-      <p>Ⓒ <a href="https://github.com/EarthyScience/FireSight?tab=Apache-2.0-1-ov-file#readme" target="_blank" rel="noreferrer">Apache License, Version 2.0</a></p>
-      <span>Contact :<a href="https://lazarusa.github.io/" target='_blank' rel="noreferrer">Lazaro Alonso</a>&
+      <span>Ⓒ <a href="https://lazarusa.github.io/" target='_blank' rel="noreferrer">Lazaro Alonso</a>&
       <a href="https://www.bgc-jena.mpg.de/person/jpoehls/2206" target='_blank' rel="noreferrer">Jeran Poehls</a>
       </span>
 
     </div>
-    <div className="small-screen-logos">
+    <div className="small-screen-logo">
       <a href="https://www.bgc-jena.mpg.de/en/bgi/home" target="_blank" rel="noreferrer">
-        <img src={logoMPI} alt="logoMPI" className="small-screen-logo" height={35}/>
+        <Image src={logoMPI} alt="logoMPI" className="small-screen-logo" height={32}/>
       </a>
     </div>
   </div>
