--- src/components/UI/Metadata.css
@@ -1,11 +1,0 @@
-.metadata{
-    position: absolute;
-    width: 250px;
-    font-size: 12px;
-    left: 10px;
-    top: 50px; 
-    background-color: rgba(255, 255, 255, 0.247);
-    padding: 10px;
-    border-radius: 10px;
-    overflow-wrap: break-word;
-}
