--- src/components/UI/Navbar.css
@@ -1,49 +1,0 @@
-
-.navbar {
-  display: flex;
-  position: fixed;
-  justify-content: space-between;
-  top: 2px;
-  left: 2px;
-  z-index: 1;
-}
-
-.navbar img {
-  max-width: 38px;
-  max-height: 38px;
-  background-color: #000000aa;
-  border-radius: 0.5rem;
-}
-
-
-.navbar ul {
-    list-style: none;
-    display:flex;
-    gap: 1rem;
-}
-
-.navbar a {
-    position: relative;
-    text-decoration: none;
-    color: var(--primary-text-color);
-  }
-  
-  .navbar a::after {
-    content: '';
-    position: absolute;
-    width: 0;
-    height: 2px;
-    bottom: 0;
-    left: 0;
-    background-color: #f5ba25;
-    transition: width .3s;
-  }
-
-.navbar a:hover {
-    text-decoration: none; /* Remove underline */
-    color:var(--tp-button-background-color-focus);
-}
-  
-.navbar a:hover::after {
-  width: 100%;
-}
