+++ src/components/UI/Navbar.tsx
@@ -1,13 +1,16 @@
+import Image from "next/image";
 import { logoSeasFire } from "@/assets/index";
-import AboutButton from "@/components/UI/AboutButton";
-import './Navbar.css'
+import { AboutButton } from "@/components/ui";
+import ThemeSwitch  from "./ThemeSwitch";
+import './css/Navbar.css'
 
 const Navbar = () => {
   return (
     <nav className="navbar">
-      <a href="https://github.com/EarthyScience/ViZarrDev" target="_blank" rel="noreferrer">
-        <img src={logoSeasFire} alt="ViZarrDev"/>
+      <a>
+        <Image src={logoSeasFire} alt="vizarrstores" />
       </a>
+      <ThemeSwitch />
       <AboutButton />
     </nav>
   );
