+++ src/components/UI/Overlays.tsx
@@ -1,12 +1,27 @@
+'use client';
+
 import { createPortal } from "react-dom";
 import { useAtomValue } from "jotai";
-import AboutModal from "@/components/UI/AboutModal";
-import {uiAtom } from "@/state";
-
-const mountElement = document.getElementById("overlays");
+import { useEffect, useState } from "react";
+import { AboutModal, uiAtom } from "@/components/ui";
 
 const Overlays = () => {
   const ui = useAtomValue(uiAtom);
-  return mountElement ? createPortal(<>{ui.modal && <AboutModal />}</>, mountElement) : null;
+  const [mounted, setMounted] = useState(false);
+
+  useEffect(() => {
+    setMounted(true);
+  }, []);
+
+  return (
+    <>
+      <div id="overlays" />
+      {mounted && createPortal(
+        <>{ui.modal && <AboutModal />}</>,
+        document.getElementById("overlays")!
+      )}
+    </>
+  );
 };
+
 export default Overlays;
