--- src/components/UI/ResizeBar.css
@@ -1,9 +1,0 @@
-.resize-bar{
-    position: absolute;
-    width: 100vw;
-    height: 10px;
-    z-index: 11;
-    cursor: n-resize;
-    background-color: transparent;
-    
-}
