+++ src/components/UI/css/AboutButton.css
@@ -1,0 +1,12 @@
+.button-about {
+    border-radius: 8px;
+    padding: 0.3em 0.5em;
+    font-size: 14px;
+    font-weight: 500;
+    font-family: inherit;
+    cursor: pointer;
+    position: fixed;
+    top: 4px;
+    right: 8px;
+    z-index: 9999;
+  }
