+++ src/components/UI/css/AboutModal.css
@@ -1,0 +1,59 @@
+.modal {
+  position: fixed;
+  top: 0;
+  left: 0px;
+  z-index: 10;
+  width: 100vw;
+  height: 100vh;
+  background-color: var(--background-modal);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  overflow: hidden;
+}
+
+.modal-content {
+  background-image: linear-gradient(62deg, var(--background-modal) 0%, var(--modal-gradient) 100%);
+  position: relative;
+  max-width: 25rem;
+  max-height: 80vh; /* Limit the max-height to 80% of viewport height */
+  margin: 0 auto;
+  padding: 1rem;
+  border-radius: 0.5rem;
+  box-shadow: 0 0 1em var(--modal-shadow);
+  overflow-y: auto; /* Enable vertical scrolling */
+  z-index: 1001;
+}
+
+.modal-content p {
+  color: var(--modal-p);
+  font-size: 14px;
+}
+
+.modal-content ul {
+  color: var(--modal-p);
+  font-size: 14px;
+}
+
+.close-btn {
+  position: absolute;
+  top: -0.9rem;
+  left: 0.1rem;
+  color: rgb(70, 70, 70);
+  font-size: 36px;
+  cursor: pointer;
+}
+
+.close-btn:hover {
+  color: var(--accent-1);
+}
+
+.modal-content a {
+  text-align: center;
+}
+
+.modal-content h1 {
+  color: rgb(128, 125, 125);
+  font-size: 1.5rem;
+  text-align: center;
+}
