+++ src/components/UI/css/Footer.css
@@ -1,0 +1,98 @@
+.footer {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    z-index: 1;
+    font-size: 13.5px;
+    position: fixed;
+    bottom: 0;
+    left: 0;
+    padding: 0.25rem 0.5rem;
+    width: 100%;
+    height: 48px;
+}
+.footer img {
+    max-height: 36px;
+}
+
+[data-theme='light'] {
+    .footer img {
+        filter: invert(0.8);
+    }
+}
+
+.footer p {
+    margin: 0;
+    flex: 1;
+    text-align: center;
+}
+.footer span {
+    margin: 0;
+    text-align: center;
+}
+.footer a {
+    position: relative;
+    text-decoration: none;
+    padding: 0 0.25rem;
+}
+
+.large-screen-logo {
+    display: block;
+}
+
+.footer a::after {
+    content: '';
+    position: absolute;
+    width: 0;
+    height: 1.25px;
+    bottom: 0;
+    left: 0;
+    background-color: rgb(205, 193, 193);
+    transition: width 0.3s;
+}
+
+
+.footer a:hover::after {
+    width: 100%;
+}
+
+.small-screen-logo {
+    display: none;
+}
+.expandable-text {
+    display: flex;
+    flex: 1;
+    white-space: nowrap;
+    justify-content: flex-end;
+    margin-right: 4px;
+}
+
+/* Responsive design for screens below 600px */
+@media (max-width: 600px) {
+    .footer {
+        display: grid;
+        grid-template-columns: auto 1fr auto;
+        align-items: center;
+        text-align: center;
+    }
+
+    .footer p {
+        text-align: center;
+        flex: unset;
+        margin: 0;
+    }
+
+    .large-screen-logo {
+        display: none;
+    }
+
+    .small-screen-logo {
+        display: flex;
+        justify-content: space-between;
+    }
+    .expandable-text {
+        display: unset;
+        margin-left: 0;
+        margin-right: 0;
+    }
+}
