+++ src/components/UI/state.tsx
@@ -1,0 +1,6 @@
+import { atom } from "jotai";
+
+export const uiAtom = atom({
+  modal: false,
+  stats: false,
+});
