--- src/main.tsx
@@ -1,10 +1,0 @@
-import * as React from 'react';
-import * as ReactDOM from 'react-dom/client'
-import App from '@/App.tsx'
-import './styles.css'
-
-ReactDOM.createRoot(document.getElementById('root')!).render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-)
