--- src/state.tsx
@@ -1,6 +1,0 @@
-import { atom } from "jotai";
-
-export const uiAtom = atom({
-  modal: false,
-  stats: false,
-});
