--- src/styles.css
@@ -1,113 +1,0 @@
-* {
-  box-sizing: border-box;
-}
-
-#root {
-  width: 100%;
-  height: 100%;
-  margin: 0;
-  padding: 0;
-  overflow: hidden;
-  font-size: 14px;
-  color: #cfcbcb;
-  --tp-base-background-color: hsla(0, 0%, 10%, 0.8);
-  --tp-base-shadow-color: hsla(0, 0%, 0%, 0.2);
-  --tp-button-background-color: hsla(0, 0%, 80%, 1);
-  --tp-button-background-color-active: hsla(0, 0%, 100%, 1);
-  --tp-button-background-color-focus: hsla(0, 0%, 95%, 1);
-  --tp-button-background-color-hover: hsla(0, 0%, 85%, 1);
-  --tp-button-foreground-color: hsla(0, 0%, 0%, 0.8);
-  --tp-container-background-color: hsla(0, 0%, 0%, 0.3);
-  --tp-container-background-color-active: hsla(0, 0%, 0%, 0.6);
-  --tp-container-background-color-focus: hsla(0, 0%, 0%, 0.5);
-  --tp-container-background-color-hover: hsla(0, 0%, 0%, 0.4);
-  --tp-container-foreground-color: hsla(0, 0%, 100%, 0.5);
-  --tp-groove-foreground-color: hsla(0, 0%, 0%, 0.2);
-  --tp-input-background-color: hsla(0, 0%, 0%, 0.3);
-  --tp-input-background-color-active: hsla(0, 0%, 0%, 0.6);
-  --tp-input-background-color-focus: hsla(0, 0%, 0%, 0.5);
-  --tp-input-background-color-hover: hsla(0, 0%, 0%, 0.4);
-  --tp-input-foreground-color: hsla(0, 0%, 100%, 0.5);
-  --tp-label-foreground-color: hsla(0, 0%, 100%, 0.5);
-  --tp-monitor-background-color: hsla(0, 0%, 0%, 0.3);
-  --tp-monitor-foreground-color: hsla(0, 0%, 100%, 0.3);
-}
-
-.canvas {
-  display: block;
-  height: 100vh;
-  width: 100vw;
-  margin: 0;
-  padding: 0;
-}
-
-.timeseries{
-  height: 100vh;
-  width: 100vw;
-  margin: 0;
-  padding: 0;
-  z-index: 10;
-  background-color: transparent;
-}
-
-body {
-  overscroll-behavior-y: none;
-  background: #151616;;
-  font-family: Inter, Avenir, Helvetica, Arial, sans-serif;
-  text-rendering: optimizeLegibility;
-  -webkit-font-smoothing: antialiased;
-  --primary-text-color: #141313;
-  --secondary-text-color: #eaeaea;
-  --accent-color1: #339ecf;
-  --accent-color2: #f38e2a;
-}
-
-
-a {
-  color: black;
-}
-
-a {
-  pointer-events: all;
-  color: black;
-  text-decoration: none;
-}
-
-svg {
-  fill: black;
-}
-
-.button-about {
-  border-radius: 8px;
-  border: 0.1rem solid transparent;
-  padding: 0.3em 0.5em;
-  font-size: 1em;
-  font-weight: 500;
-  font-family: inherit;
-  background-color: var(--tp-base-background-color);
-  cursor: pointer;
-  transition: border-color 0.25s;
-  position: fixed;
-  top: 4px;
-  right: 2px;
-  z-index: 9999;
-}
-.button-about:hover {
-  border-color: var(--tp-input-foreground-color);
-}
-.button-about:focus,
-.button-about:focus-visible {
-  outline: 4px auto -webkit-focus-ring-color;
-  outline: 4px auto Highlight;
-}
-.button-about a {
-  color: var(--tp-button-background-color-focus); /* Change to the color you want */
-}
-.button-about a:active {
-  color: #c7c2c2; /* Change to the color you want */
-}
-
-.plot-canvas {
-  border-top: 1px solid #333;
-  box-shadow: 0 -2px 10px #00000080;
-}
