--- src/vite-env.d.ts
@@ -1,6 +1,0 @@
-/// <reference types="vite/client" />
-
-declare module '*.glsl' {
-    const value: string;
-    export default value;
-  }
